Renamed themes.py to styling.py and propogated changes throughout
